feat: criando e estilizando a home. criando roteamento entre as páginas.
