+++ best-browser-games/public/pagina-de-destino-de-jogos-de-neon-desenhada-a-mao/Fonts.txt
@@ -1,0 +1,9 @@
+Fonts used in this design:
+
+Font: Inter
+Author: Rasmus Andersson
+Url: https://fonts.google.com/specimen/Inter
+
+Font: Press Start 2P
+Author: CodeMan38
+Url: https://fonts.google.com/specimen/Press+Start+2P
