--- best-browser-games/src/components/CadastroForm.css
@@ -1,50 +1,0 @@
-body {
-    font-family: 'Arial', sans-serif;
-    background-image: url('../../public/photocall-de-jogos-de-neon-desenhado-a-mao//8088488.jpg');
-    background-size: cover; 
-    background-position: center; 
-    margin: 10%;
-    padding: 0;
-    display: flex;
-    height: 100vh;
-    justify-content: center; 
-    align-items: center; 
-  }
-  
-  form {
-    width: 300px;
-    margin: 20px auto;
-  }
-  
-  label {
-    display: block;
-    margin-bottom: 10px;
-    color: #b1d22d;
-  }
-  
-  input {
-    width: 100%;
-    padding: 8px;
-    margin-bottom: 15px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    box-sizing: border-box;
-  }
-  
-  input:focus {
-    border-color: #ffc160;
-  }
-  
-  button {
-    padding: 10px 20px;
-    background-color: #afd32c;
-    color: #000;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-  }
-  
-  button:hover {
-    background-color: #ffc160;
-  }
-
