+++ best-browser-games/src/index.css
@@ -1,5 +1,9 @@
 body {
   margin: 0;
+  padding: 0;
+  background-size: cover;
+  background-position: center;
+  background-size: 100% auto;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
